Update README.md
This is an example of how you can use events, animations, triggers and effects to create a more complex lens using lens studio. The C++ code creates a class called "MouthOpenColorChange" which inherits from the SnapObject class provided by Lens Studio. 
This class has 3 private variables r, g, and b which represent the color of the object. Also, it has current_animation variable that will be used to play different animations.

When the lens is started, the OnStart() function is called. In this function, the class registers for the
"MOUTH_OPEN", "BUTTON_PRESS", and "SLIDER_CHANGE" events using the GetEventDispatcher()->RegisterEvent("EVENT_NAME", this) function. The "MOUTH_OPEN" event is triggered when the user's mouth is opened. The "BUTTON_PRESS" event is triggered when a button is pressed by the user and "SLIDER_CHANGE" event is triggered when the user changes the value of a slider.

When the "MOUTH_OPEN" event is triggered, the OnEvent(const Event &event) function is called. Inside this function, an if statement checks if the event name is "MOUTH_OPEN". If it is, the rust_change_color(r, g, b) and rust_play_animation(current_animation) functions are called. The r, g, b are passed as arguments to the rust_change_color function, which will change the color of the object. The current_animation variable is passed to the rust_play_animation function, which will play the corresponding animation on the object. The r, g, b values will be incremented by 85 each time the event is triggered, to change the color of the object.

Then, there is an else if statement that checks if the event name is "BUTTON_PRESS". If it is, the code inside the block will handle the button press event. Example would be incrementing the current_animation variable so that the next time the mouth is open, the next animation will play.

Another else if statement that checks if the event name is "SLIDER_CHANGE". If it is, the code inside the block will handle the slider change event. Example would be getting the value of size from the event, and then changing the scale of the object with this value.

The rust code defines two functions, rust_change_color and rust_play_animation. The rust_change_color function takes 3 arguments r, g, and b, which represent the color of the object. The rust function will execute the code to change the color of an object in the lens, using the r, g, b values passed from the C++ wrapper. The rust_play_animation function takes one argument, animation, that represents the animation that needs to be played. It will execute the code to play the corresponding animation on the object in the lens, using the animation value passed from the C++ wrapper.

Tracking objects: You could use the Object Tracker feature to track specific objects in the camera view, such as a person's face or a bottle. You could then use this information to control the position and orientation of objects in the lens.

Face filters: You could use the Face Tracker feature to detect and track specific facial features, such as the eyes, nose, or mouth. You could then use this information to create interactive effects, such as changing the color of an object based on the user's eye movement or creating a 3D model that follows the user's face.

3D models: You could use the Model Importer to import 3D models into your lens and use the 3D Transformation tool to position and rotate them in 3D space. You could also create animations for the models and use the Animation Track to control the timing and speed of the animations based on user interactions.